Fixes problems with WORMStrategy use in HA.  Specifically addresses live message compression and HALog playback for both compressed and uncompressed content required in Resync.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/READ_CACHE@7180 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/ha/pipeline/HAReceiveService.java
@@ -1078,7 +1078,7 @@ public class HAReceiveService<M extends IHAWriteMessageBase> extends Thread {
 
 			if (message.getChk() != (int) chk.getValue()) {
                 throw new ChecksumError("msg=" + message.toString()
-                        + ", actual=" + chk.getValue());
+                        + ", actual=" + (int) chk.getValue());
             }
 
             if (callback != null) {
